Make the scramble count a log slider: 10 to 10 billion
Slider stops are powers of ten (10^1..10^10). Scramble's count
parameter widens from int to double (as the 2008 GetDigitBoxValue
was) since 10^10 exceeds int32, the completion log prints the count
with %.0f, and the JS moves display detects uint32 counter wraps so
10-billion-move runs read correctly. Verified a 3.7e9 request is
accepted un-truncated.

+++ web/wasm/api.cpp
@@ -53,7 +53,7 @@ EMSCRIPTEN_KEEPALIVE int eng_save_ppm( const char *path )
 	return g_Engine.SavePPM( path ) ? 1 : 0;
 }
 
-EMSCRIPTEN_KEEPALIVE void eng_scramble( int n, int swirl, int direction )
+EMSCRIPTEN_KEEPALIVE void eng_scramble( double n, int swirl, int direction )
 {
 	g_Engine.Scramble( n, swirl != 0, direction != 0 );
 }
